Update 2021-07-30-minisummit-agenda.md
I'm a dork, and did not compose my abstract for public consumption in a conference agenda. I know, I should have known that. Edited to not be the lone abstract written in the 3rd person, to have it read a touch smoother, and to tone-down what read as puffery salesmanship (which, well, it was—to get the talk accepted).